add an ugly workaround so that repo_write knows how to deal with filelist extensions
See BUGS file. Without that patch, repo_write wrote both the filelist in the
primary data and the full filelists, resulting in a broken solv file. Now, the
filelist from the primary file will not be written.
Note that this is just a bad workaround and should be fixed completely different.
See BUGS file.