Implement libsolv Tcl bindings
Add the necessary build system changes and modifications in the SIWG
interface definition file to enable generation of Tcl bindings. The
build system generates and installs a Tcl module that can used from
tclsh using
  % package require solv
Constants and operations are available in the solv:: namespace, e.g.
  % puts $solv::Solver_POLICY_ILLEGAL_DOWNGRADE
will echo "1".

Rudimentary testing has been done, but this should be considered
experimental for now. We will fix problems as we encounter them.

Signed-off-by: Clemens Lang <neverpanic@gmail.com>