Fix last commit
We can't add the multiversion orphan to the update rules, as that
would mean we no longer enforce an update to one of the repo
packages. So we now just detect the multiversion orphan and do not
add a dup rule.

Note that this is still not correct, as we don't enforce that the
orphan is not erased in keep_orphan mode. In that case we really
need two independent rules. I'm currently not sure where to put
the extra rule...