Overshoot size for large memory chunks in solv_extend
We now dynamically round up the size to some ppower of two. This
brings down the number of mremap calls fro 96163 to 675 when
converting Fedora22's Everything repo, which makes no real difference
on my system but seems to help a lot if this is done in a Xen
guest. See libsolv issue#111 on github.