Strore decision reason in decisionq_reason
This makes the code simpler and more flexible. Do away with
the old way of deducting the reason from the decision index,
instead store the reason in the decision process.

Note that the decisionq_reason is indexed by level, as all
the other decisions for a level are because of unit rules.
Thus the reason queue is much smaller than the decision queue.