Always create dup rules for distupgrade jobs
We used to use update rules for "distupgrade all packages".
This is a tiny bit faster and needs less memory, but the downside
is that for problems there only is the "get rid of the update rule"
solution, but not the "disable dup mode for this package" solution.

So simplify things by always creating dup rules for dup mode.