Rework fetch phase in fileconflict handling
We now handle all files of one package in one go instead of
iterating over every file. In most cases this does not make
a difference, as the pruned candidate list is very small. But
it can be big if two packages with the same content get
installed, e.g. two kernels with same version but different
rebuild count.