Work around rpm bug if HEADERIMPORT_FAST is used
With HEADERIMPORT_FAST, rpm uses the difference between offsets
to determine the entry length. Unfortunately, the length is
used as entry count for RPM_BIN_TYPE, so the cound may end up
to big. So trim it again to match the md5 length.

Thanks to Oleg Girko <ol@infoserver.lv> for investigating this.