Add const for second variable
The second variable is used as a source for copy and it is not modified
therefore it should be marked as s const. Then it is more easy to identify what
is source and what is target and also it is clear that the source variable will
be not modify by the function.

Additionally it allows to use const map or queue to create a copy or other
operations.