Refactor arch handling
lastarch is no longer the last element of id2arch, but now the
size of the array. This means that lastarch == 0 now reliable means
that no architecture was set.

To be compatible to old versions we over-allocate id2arch so that
id2arch[lastarch] is always zero.

There's also a new pool_arch2score function that abstracts accessing
the id2arch array.