Fix library type of tcl binding
Building the tcl binding as a SHARED type library means that an INSTALL of
that target is treated as a RUNTIME target on DLL platforms, which currently
has no DESTINATION provided.

Change the type to MODULE (the same as other bindings), as an INSTALL of a
target of that type uses the LIBRARY target on all platforms.