Add REPOSITORY_FILTEREDFILELIST id and filelist type
REPOSITORY_FILTEREDFILELIST is used to mark that the filelist
in the repodata is filtered with some filter.

The commit also adds a filelisttype element to the repodata
and sets it to filtered if REPOSITORY_FILTEREDFILELIST is set.
It also sets the filelisttype to extension if the repodata
has a filelist and was generated by extending solvables.

We'll use the filelisttype later to decide if we should use
the filtered filelist when searching for files.