Add support for modular updateinfo.xml data
Module advisory data can specify which advisory is applicable for which
modular stream. Module information can be absent or there could be
multiple entries.

If there is a module entry in advisory and none of modules in advisory
is enabled on the system, advisory is not relevant for the system and
should be ignored.