Not considered excluded packages as a best candidate
In some work-flows we can end up in selection of a best candidate with
a package that is not in considered map, therefore it cannot be
installed. Removal of those packages as a first prune process solves
the issue. If all of solvables are not in considered map it removes all
of them, because it is required for update jobs, but not harmful for
install.

https://bugzilla.redhat.com/show_bug.cgi?id=1677583
https://bugzilla.redhat.com/show_bug.cgi?id=1678911
https://bugzilla.redhat.com/show_bug.cgi?id=1673851