Add POOL_FLAG_WHATPROVIDESWITHDISABLED pool flag
This makes libsolv keep disabled packages (i.e. packages
not in the considered map) in the whatprovides hash.

It is useful if you want to modify the considered map after
the whatprovides hash is generated.

Note that using the flag will change the reported problems
as disabled packages are now seen by the solver but
considered uninstallable.