Add solv_fmemopen() function
The solv_xfopen_buf() function we have has the disadvantage
that it updates the buffer pointer/length when reading/writing.
This can easily lead to writing to locations that went out of
scope, like with our usage in testcase.c.

So solv_fmemopen() removes that indirection and directly takes
the buffer pointer and length as arguments. Only reading is
currently supported.