Save memory in repo_updateinfoxml by not interleaving repo_addflexarray calls
Mixing repo_addflexarray calls will make the code moving the arrays
in the idarraydata all the time as it cannot append to the arrays.
So first collect the array contents of the collection flexarray and
then add it in one go.

This is based on pull request #533 by Aleš Matěj <amatej@redhat.com>.